fixed create page & component validation

+++ cmds/create-page.js
@@ -3,24 +3,28 @@ const path = require("path");
 const { capitalizeFirstLetter } = require("../cmds/utils/helpers");
 
 const createPage = argv => {
-  const { pageName } = argv;
-  const pagesTargetPath = `${process.cwd()}/src/pages/${pageName}`;
+  if (fs.existsSync(path.resolve(process.cwd(), "porsea.config.js"))) {
+    const { pageName } = argv;
+    const pagesTargetPath = `${process.cwd()}/src/pages/${pageName}`;
 
-  fs.mkdirSync(pagesTargetPath);
-  console.log("Created a New Page: " + "'" + pageName + "'");
+    fs.mkdirSync(pagesTargetPath);
+    console.log("Created a New Page: " + "'" + pageName + "'");
 
-  const baseTemplatePagePath = path.join(
-    __dirname,
-    "../templates/common/page.txt"
-  );
+    const baseTemplatePagePath = path.join(
+      __dirname,
+      "../templates/common/page.txt"
+    );
 
-  fs.readFile(baseTemplatePagePath, (_, data) => {
-    let replacePageName = data
-      .toString()
-      .replace(/{{ PageName }}/g, capitalizeFirstLetter(pageName))
-      .replace(/{{ PagePath }}/g, pageName);
-    fs.writeFileSync(pagesTargetPath + "/index.js", replacePageName);
-  });
+    fs.readFile(baseTemplatePagePath, (_, data) => {
+      let replacePageName = data
+        .toString()
+        .replace(/{{ PageName }}/g, capitalizeFirstLetter(pageName))
+        .replace(/{{ PagePath }}/g, pageName);
+      fs.writeFileSync(pagesTargetPath + "/index.js", replacePageName);
+    });
+  } else {
+    console.log("This is NOT a Porsea Project!");
+  }
 };
 
 module.exports = createPage;
